+++ src-tests/Selenium/SeleniumServerAdapterTest.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Lmc\Steward\Selenium;
 
@@ -14,12 +14,12 @@ class SeleniumServerAdapterTest extends TestCase
     /** @var string */
     protected $serverUrl = 'http://selenium.local:1337';
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->adapter = new SeleniumServerAdapter($this->serverUrl);
     }
 
-    public function testShouldGetParsedServerUrlParts()
+    public function testShouldGetParsedServerUrlParts(): void
     {
         $adapter = new SeleniumServerAdapter('https://user:pass@host:1337/selenium?foo=bar');
 
@@ -39,10 +39,8 @@ class SeleniumServerAdapterTest extends TestCase
 
     /**
      * @dataProvider provideServerUrl
-     * @param string $providedServerUrl
-     * @param string $expectedServerUrl
      */
-    public function testShouldGetServerUrlWithDefaultPortIfNeeded($providedServerUrl, $expectedServerUrl)
+    public function testShouldGetServerUrl(string $providedServerUrl, string $expectedServerUrl): void
     {
         $adapter = new SeleniumServerAdapter($providedServerUrl);
 
@@ -52,10 +50,11 @@ class SeleniumServerAdapterTest extends TestCase
     /**
      * @return array[]
      */
-    public function provideServerUrl()
+    public function provideServerUrl(): array
     {
         return [
-            'protocol, host and port' => ['http://foo:80', 'http://foo:80'],
+            'protocol, host and port => should not be changed' => ['http://foo:80', 'http://foo:80'],
+            'trailing slash should be removed' => ['http://localhost:4444/wd/hub/', 'http://localhost:4444/wd/hub'],
             'local URL with specified port should keep it' => ['http://localhost:4444', 'http://localhost:4444'],
             'local URL without port should use 4444' => ['http://localhost', 'http://localhost:4444'],
             'cloud service URL with port should keep it' =>
@@ -67,19 +66,18 @@ class SeleniumServerAdapterTest extends TestCase
             'TestingBot cloud service without port should use 80' =>
                 ['http://foo:bar@hub.testingbot.com', 'http://foo:bar@hub.testingbot.com:80'],
             'non-cloud URL without port should use 4444' => ['http://foo.com', 'http://foo.com:4444'],
-            'username and host should get deault port' => ['http://user@foo', 'http://user@foo:4444'],
-            'all parts' =>
-                ['https://user:pass@host:1337/selenium?foo=bar', 'https://user:pass@host:1337/selenium?foo=bar'],
+            'username and host should get default port' => ['http://user@foo', 'http://user@foo:4444'],
+            'all url parts' =>
+                ['https://user:pass@host:1337/wd/hub?foo=bar', 'https://user:pass@host:1337/wd/hub?foo=bar'],
             'all parts expects port should get default port' =>
-                ['https://user:pass@host/selenium?foo=bar', 'https://user:pass@host:4444/selenium?foo=bar'],
+                ['https://user:pass@host/wd/hub?foo=bar', 'https://user:pass@host:4444/wd/hub?foo=bar'],
         ];
     }
 
     /**
      * @dataProvider provideInvalidServerUrl
-     * @param string $serverUrl
      */
-    public function testShouldThrowExceptionIfInvalidServerUrlIsGiven($serverUrl)
+    public function testShouldThrowExceptionIfInvalidServerUrlIsGiven(string $serverUrl): void
     {
         $this->expectException(\RuntimeException::class);
         $this->expectExceptionMessage('Provided Selenium server URL "' . $serverUrl . '" is invalid');
@@ -90,7 +88,7 @@ class SeleniumServerAdapterTest extends TestCase
     /**
      * @return array[]
      */
-    public function provideInvalidServerUrl()
+    public function provideInvalidServerUrl(): array
     {
         return [
             ['http://'],
@@ -99,39 +97,7 @@ class SeleniumServerAdapterTest extends TestCase
         ];
     }
 
-    /**
-     * @dataProvider provideServerUrlWithHubEndpoint
-     * @param string $providedServerUrl
-     * @param string $expectedServerUrl
-     */
-    public function testShouldTrimHubEndpointFromTheServerUrl($providedServerUrl, $expectedServerUrl)
-    {
-        $adapter = new SeleniumServerAdapter($providedServerUrl);
-
-        $this->assertSame($expectedServerUrl, $adapter->getServerUrl());
-    }
-
-    /**
-     * @return array[]
-     */
-    public function provideServerUrlWithHubEndpoint()
-    {
-        return [
-            'No endpoint, URL with port' => ['http://localhost:4444', 'http://localhost:4444'],
-            'No endpoint, URL without port' => ['http://localhost', 'http://localhost:4444'],
-            'Endpoint defined, with port' => ['http://localhost:4444/wd/hub', 'http://localhost:4444'],
-            'Endpoint defined, without port' => ['http://localhost/wd/hub', 'http://localhost:4444'],
-            'Endpoint with trailing slash & port' => ['http://localhost:4444/wd/hub/', 'http://localhost:4444'],
-            'Endpoint with trailing slash & without port' => ['http://localhost/wd/hub/', 'http://localhost:4444'],
-            'Not a hub endpoint should be kept' => ['http://localhost/foo/bar', 'http://localhost:4444/foo/bar'],
-            'Not a hub endpoint (with port) should be kept' => [
-                'http://localhost:1337/foo/bar',
-                'http://localhost:1337/foo/bar',
-            ],
-        ];
-    }
-
-    public function testShouldReturnTrueIfUrlIsAccessible()
+    public function testShouldReturnTrueIfUrlIsAccessible(): void
     {
         $dummyResource = fopen(__FILE__, 'r');
 
@@ -143,7 +109,7 @@ class SeleniumServerAdapterTest extends TestCase
         $this->assertTrue($this->adapter->isAccessible());
     }
 
-    public function testShouldReturnFalseIfConnectionIsRefused()
+    public function testShouldReturnFalseIfConnectionIsRefused(): void
     {
         $fsockopenMock = $this->getFunctionMock(__NAMESPACE__, 'fsockopen');
         $fsockopenMock->expects($this->once())
@@ -157,71 +123,92 @@ class SeleniumServerAdapterTest extends TestCase
         $this->assertEquals('Connection refused', $this->adapter->getLastError());
     }
 
-    public function testShouldReturnFalseIfTheServerDoesNotRespondToStatusUrl()
+    public function testShouldReturnFalseIfTheServerDoesNotRespondToStatusUrl(): void
     {
         $fileGetContentsMock = $this->getFunctionMock(__NAMESPACE__, 'file_get_contents');
         $fileGetContentsMock->expects($this->once())
-            ->with($this->serverUrl . '/wd/hub/status')
+            ->with($this->serverUrl . '/status')
             ->willReturn(false);
 
         $this->assertFalse($this->adapter->isSeleniumServer());
         $this->assertEquals('error reading server response', $this->adapter->getLastError());
     }
 
-    public function testShouldReturnJsonErrorDescriptionIfTheServerResponseIsNotJson()
+    public function testShouldDetectNotReadySeleniumServer(): void
+    {
+        $notReadyResponse = file_get_contents(__DIR__ . '/Fixtures/response-standalone-w3c-not-ready.json');
+        $fileGetContentsMock = $this->getFunctionMock(__NAMESPACE__, 'file_get_contents');
+        $fileGetContentsMock->expects($this->once())
+            ->with($this->serverUrl . '/status')
+            ->willReturn($notReadyResponse);
+
+        $this->assertFalse($this->adapter->isSeleniumServer());
+        $this->assertEquals('server is not ready ("No spare hub capacity")', $this->adapter->getLastError());
+    }
+
+    public function testShouldReturnJsonErrorDescriptionIfTheServerResponseIsNotJson(): void
     {
         $fileGetContentsMock = $this->getFunctionMock(__NAMESPACE__, 'file_get_contents');
         $fileGetContentsMock->expects($this->once())
-            ->with($this->serverUrl . '/wd/hub/status')
+            ->with($this->serverUrl . '/status')
             ->willReturn('THIS IS NOT JSON');
 
         $this->assertFalse($this->adapter->isSeleniumServer());
         $this->assertRegExp('/^error parsing server JSON response \(.+\)$/', $this->adapter->getLastError());
     }
 
-    public function testShouldReturnTrueIfServerRespondsWithJsonInNoGridMode()
+    /**
+     * @dataProvider provideValidStatusResponses
+     */
+    public function testShouldDetectSeleniumServer(string $responseFixtureFile, string $expectedCloudService): void
     {
-        $response = file_get_contents(__DIR__ . '/Fixtures/response-standalone.json');
+        $response = file_get_contents($responseFixtureFile);
         $fileGetContentsMock = $this->getFunctionMock(__NAMESPACE__, 'file_get_contents');
         $fileGetContentsMock->expects($this->once())
-            ->with($this->serverUrl . '/wd/hub/status')
+            ->with($this->serverUrl . '/status')
             ->willReturn($response);
 
         $this->assertTrue($this->adapter->isSeleniumServer());
-        $this->assertEmpty($this->adapter->getCloudService());
+        $this->assertSame($this->adapter->getCloudService(), $expectedCloudService);
         $this->assertEmpty($this->adapter->getLastError());
     }
 
-    public function testShouldReturnTrueIfServerRespondsWithJsonInGridMode()
+    /**
+     * @return array[]
+     */
+    public function provideValidStatusResponses(): array
     {
-        $response = file_get_contents(__DIR__ . '/Fixtures/response-grid.json');
-        $fileGetContentsMock = $this->getFunctionMock(__NAMESPACE__, 'file_get_contents');
-        $fileGetContentsMock->expects($this->once())
-            ->with($this->serverUrl . '/wd/hub/status')
-            ->willReturn($response);
-
-        $this->assertTrue($this->adapter->isSeleniumServer());
-        $this->assertEmpty($this->adapter->getCloudService());
-        $this->assertEmpty($this->adapter->getLastError());
+        return [
+            // $responseData, $expectedCloudService
+            'standalone W3C server (in ready state)' => [__DIR__ . '/Fixtures/response-standalone-w3c-ready.json', ''],
+            'standalone server v2' => [__DIR__ . '/Fixtures/response-standalone-v2.json', ''],
+            'standalone local grid v2' => [__DIR__ . '/Fixtures/response-standalone-hub-v2.json', ''],
+            'Sauce Labs cloud' =>
+                [__DIR__ . '/Fixtures/response-saucelabs.json', SeleniumServerAdapter::CLOUD_SERVICE_SAUCELABS],
+            'BrowserStack cloud' =>
+                [__DIR__ . '/Fixtures/response-browserstack.json', SeleniumServerAdapter::CLOUD_SERVICE_BROWSERSTACK],
+            'TestingBot cloud' =>
+                [__DIR__ . '/Fixtures/response-testingbot.json', SeleniumServerAdapter::CLOUD_SERVICE_TESTINGBOT],
+        ];
     }
 
-    public function testShouldConnectToTheServerOnlyOnceWhenAttemptingToGetCloudServiceName()
+    public function testShouldConnectToTheServerOnlyOnceWhenAttemptingToGetCloudServiceName(): void
     {
-        $response = file_get_contents(__DIR__ . '/Fixtures/response-standalone.json');
+        $response = file_get_contents(__DIR__ . '/Fixtures/response-standalone-v2.json');
         $fileGetContentsMock = $this->getFunctionMock(__NAMESPACE__, 'file_get_contents');
         $fileGetContentsMock->expects($this->once())
-            ->with($this->serverUrl . '/wd/hub/status')
+            ->with($this->serverUrl . '/status')
             ->willReturn($response);
 
         $this->assertEmpty($this->adapter->getCloudService());
         $this->assertEmpty($this->adapter->getCloudService());
     }
 
-    public function testShouldThrowExceptionWhenGettingCloudServiceNameButTheServerResponseIsInvalid()
+    public function testShouldThrowExceptionWhenGettingCloudServiceNameButTheServerResponseIsInvalid(): void
     {
         $fileGetContentsMock = $this->getFunctionMock(__NAMESPACE__, 'file_get_contents');
         $fileGetContentsMock->expects($this->once())
-            ->with($this->serverUrl . '/wd/hub/status')
+            ->with($this->serverUrl . '/status')
             ->willReturn('THIS IS NOT JSON');
 
         $this->expectException(\RuntimeException::class);
@@ -233,50 +220,11 @@ class SeleniumServerAdapterTest extends TestCase
     }
 
     /**
-     * @dataProvider provideCloudServiceResponse
-     * @param string $responseData
-     * @param string $expectedCloudService
-     */
-    public function testShouldDetectCloudService($responseData, $expectedCloudService)
-    {
-        $adapter = new SeleniumServerAdapter('http://such.cloud:80');
-        $fileGetContentsMock = $this->getFunctionMock(__NAMESPACE__, 'file_get_contents');
-        $fileGetContentsMock->expects($this->once())
-            ->with('http://such.cloud:80/wd/hub/status')
-            ->willReturn($responseData);
-
-        $this->assertSame($expectedCloudService, $adapter->getCloudService());
-    }
-
-    /**
-     * @return array[]
-     */
-    public function provideCloudServiceResponse()
-    {
-        $responseSauceLabs = file_get_contents(__DIR__ . '/Fixtures/response-saucelabs.json');
-        $responseBrowserStack = file_get_contents(__DIR__ . '/Fixtures/response-browserstack.json');
-        $responseTestingBot = file_get_contents(__DIR__ . '/Fixtures/response-testingbot.json');
-        $responseStandalone = file_get_contents(__DIR__ . '/Fixtures/response-standalone.json');
-        $responseLocalGrid = file_get_contents(__DIR__ . '/Fixtures/response-grid.json');
-
-        return [
-            // $responseData, $expectedCloudService
-            'Sauce Labs' => [$responseSauceLabs, SeleniumServerAdapter::CLOUD_SERVICE_SAUCELABS],
-            'BrowserStack' => [$responseBrowserStack, SeleniumServerAdapter::CLOUD_SERVICE_BROWSERSTACK],
-            'TestingBot' => [$responseTestingBot, SeleniumServerAdapter::CLOUD_SERVICE_TESTINGBOT],
-            'non-cloud local standalone server' => [$responseStandalone, ''],
-            'non-cloud local grid' => [$responseLocalGrid, ''],
-        ];
-    }
-
-    /**
      * @dataProvider provideSessionExecutorResponse
-     * @param string $responseData
-     * @param string $expectedSessionExecutor
      */
-    public function testShouldGetSessionExecutor($responseData, $expectedSessionExecutor)
+    public function testShouldGetSessionExecutor(string $responseData, string $expectedSessionExecutor): void
     {
-        $adapter = new SeleniumServerAdapter('http://127.0.0.1:4444');
+        $adapter = new SeleniumServerAdapter('http://127.0.0.1:4444/wd/hub');
         $fileGetContentsMock = $this->getFunctionMock(__NAMESPACE__, 'file_get_contents');
         $fileGetContentsMock->expects($this->once())
             ->with('http://127.0.0.1:4444/grid/api/testsession?session=4f1bebc2-667e-4b99-b16a-ff36221a20b3')
@@ -291,7 +239,7 @@ class SeleniumServerAdapterTest extends TestCase
     /**
      * @return array[]
      */
-    public function provideSessionExecutorResponse()
+    public function provideSessionExecutorResponse(): array
     {
         $responseExecutorFound = file_get_contents(__DIR__ . '/Fixtures/testsession-found.json');
         $responseExecutorNotFound = file_get_contents(__DIR__ . '/Fixtures/testsession-not-found.json');